processor/all_requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==5.0.1
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
commonmark==0.9.1
contourpy==1.1.1
cycler==0.12.0
docutils==0.19
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.43.0
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
imageio==2.31.4
importlib-metadata==5.0.0
importlib-resources==6.1.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.11.0
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==9.0.0
networkx==3.1
nibabel==5.1.0
numpy==1.26.0
opencv-python==4.5.3.56
outcome==1.2.0
packaging==23.1
Pillow==10.0.1
pkginfo==1.8.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.13.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyWavelets==1.4.1
readme-renderer==37.3
requests==2.31.0
requests-toolbelt==0.10.1
rich==12.6.0
rsa==4.9
scikit-image==0.21.0
scipy==1.11.3
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tifffile==2023.9.26
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
twine==4.0.1
uritemplate==4.1.1
urllib3==2.0.5
webencodings==0.5.1
Werkzeug==3.0.0
wsproto==1.2.0
zipp==3.10.0
